refactor and fix ui
