+++ packages/cli/src/ui/hooks/useAutoAcceptIndicator.ts
@@ -32,19 +32,10 @@ export function useAutoAcceptIndicator({
           ? ApprovalMode.DEFAULT
           : ApprovalMode.YOLO;
     } else if (key.tab && key.shift) {
-      const currentMode = config.getApprovalMode();
-      switch (currentMode) {
-        case ApprovalMode.DEFAULT:
-          nextApprovalMode = ApprovalMode.AUTO_EDIT;
-          break;
-        case ApprovalMode.AUTO_EDIT:
-          nextApprovalMode = ApprovalMode.PLAN;
-          break;
-        case ApprovalMode.PLAN:
-        default:
-          nextApprovalMode = ApprovalMode.DEFAULT;
-          break;
-      }
+      nextApprovalMode =
+        config.getApprovalMode() === ApprovalMode.AUTO_EDIT
+          ? ApprovalMode.DEFAULT
+          : ApprovalMode.AUTO_EDIT;
     }
 
     if (nextApprovalMode) {
