+++ packages/core/src/core/prompts.ts
@@ -13,7 +13,8 @@ import { GlobTool } from '../tools/glob.js';
 import { GrepTool } from '../tools/grep.js';
 import { ReadFileTool } from '../tools/read-file.js';
 import { ReadManyFilesTool } from '../tools/read-many-files.js';
-import { ReadPlanTool, WritePlanItemTool, WritePlanTool } from '../tools/plan.js';
+import { WritePlanTool } from '../tools/plan-write.js';
+import { ReadPlanTool } from '../tools/plan-read.js';
 import { ShellTool } from '../tools/shell.js';
 import { WriteFileTool } from '../tools/write-file.js';
 import process from 'node:process';
